Internal functions for predict.fitmodel function.

Internal functions used for prediction. EstChat: estimates concentration. BlockSigma: estimates the variance-covariance matrix. CalcSE: calls on C code to calculate the standard error. ConvertFacts: Internal function that determines the conversion constant to use for loads and flow.Conversion to Million tonnes: C = mg/L Q = m^3/s (K.loads) mg/l x m3/1000 = 1kg & 1kg = 0.001 t Conversion to millions of ML (K.flow). EstLoadYS: Function used in predict.fitmodel to estimate the load based on predicted concentrations from predicted hourly flow records. EstLoad: Function called by EstLoadYS ExtractDailyLoads: extracts the daily loads. Intervals: calculates confidence intervals.

Intervals(est, pvalue)

CalcSE(Xdesign, LhatC, alpha1, alpha2, beta1, rho, W, sig2)

BlockSigma(len, Xdesign, W)

ConvertFacts(samp.unit)

EstChat(Qdat, modobj, modobjfix)

EstLoad(
  CQdat,
  Qdat,
  Xdesign,
  W,
  sig2,
  rho,
  alpha1,
  alpha2,
  beta1,
  samp.unit,
  se = TRUE,
  Blocklen = 5000
)

EstLoadYS(
  CQdat,
  Qdat,
  concY,
  modobj,
  samp.unit = samp.unit,
  alpha1,
  alpha2,
  se = TRUE,
  Blocklen = 5000
)

ExtractDailyLoads(Date, Xdesign, Lhat, LhatC, Qhat, error, modparms)

LoadPlot(x, Conc, scale = "Mt")

Arguments

est

Estimation of response from model

pvalue

pvalue

Xdesign

Design matrix from model

LhatC

Estimated concentrations

alpha1

error in flows

alpha2

error in flows

beta1

beta1

rho

correlation term

W

W matrix

sig2

error

len

len

samp.unit

sampling unit

Qdat

Qdat

modobj

modobj

modobjfix

modobjfix

CQdat

CQdat

se

se

Blocklen

Blocklen

concY

concY

Date

Date

Lhat

Lhat

Qhat

Qhat

error

error

modparms

modparms

x

x

Conc

Concentration

scale

scale