Internal functions to create a regularized flow dataset using either a smoothing spline or rainfall record using Random Forests.

Creates a regularised flow dataset using a smoothing spline or rainfall record using Random Forests. Predictions produced from the smoothing spline approach are checked to ensure predictions do not go outside the range of the raw data. If this occurs for any year a warning is issued for the user to check.

features(series, dateTime)

regularizeFlowSS(Q, Qreg)

regularizeFlowR(rainfall, dateTimeRainfall, flow, dateTimeFlow)

FitSmQ(Q, Qreg, method = "monoH.FC")

Arguments

series

series

dateTime

dateTime

Q

Q

Qreg

Qreg

rainfall

rainfall

dateTimeRainfall

dateTimeRainfall

flow

flow

dateTimeFlow

dateTimeFlow

method

spline method for interpolation