Produces diagnostic plots of a fitted loads model

S3 method for evaluating diagnostics of a fitted LRE object

Produces a series of standard suite of diagnostics for examining the fit of a loads model.

diagnostic(object)

# S3 method for class 'fitmodel'
diagnostic(object)

Arguments

object

object of class "fitmodel" resulting from the FitModel function.

Value

Produces the following diagnostics plots: residuals versus predicted points should appear randomly distributed along the line at zero. normal Quantile-Quantile (Q-Q) plot compares the quantiles from the theoretical distribution (i.e. normal) to that observed. If the 2 distributions are similar, then the points in the Q-Q plot will approximately lie on the line. Som deviation at the tails is acceptable. histogram of residuals should appear roughly symmetric and bell shaped. No skewness (elongation of the tails) or kurtosis (thickening of the tails) should be present. predicted versus observed points should appear close to the diagonal 0-1 line. predict versus obsered (WY) plots of predicted versus observed for each water year. residuals versus predicted (WY) plots of of residuals versus predicted for each water year.

@importFrom "stats" "resid" @importFrom "stats" "lag" @importFrom "mgcv" "predict.gam" @importFrom "nlme" "predict"

@references Kuhnert, P.M., Henderson, B.L., Lewis, S.E., Bainbridge, Z.T., Wilkinson, S.N. and Brodie, J.E. (2012) Quantifying total suspended sediment export from the Burdekin River catchment using the loads regression estimator tool, Water Resources Research, 48, W04533,doi:10.1029/2011WR011080.

@seealso FitModel, qqnorm,, qqline

@keywords loads, diagnostics

Details

Series of diagnostic plots for examining model fit.