Internal functions for predict.fitmodel function.

Internal functions used for prediction. EstChat: estimates concentration. BlockSigma: estimates the variance-covariance matrix. CalcSE: calls on C code to calculate the standard error. ConvertFacts: Internal function that determines the conversion constant to use for loads and flow.Conversion to Million tonnes: C = mg/L Q = m^3/s (K.loads) mg/l x m3/1000 = 1kg & 1kg = 0.001 t Conversion to millions of ML (K.flow). EstLoadYS: Function used in predict.fitmodel to estimate the load based on predicted concentrations from predicted hourly flow records. EstLoad: Function called by EstLoadYS ExtractDailyLoads: extracts the daily loads. Intervals: calculates confidence intervals.

BlockSigma(len, Xdesign, W)

CalcSE(Xdesign, LhatC, alpha1, alpha2, beta1, rho, W, sig2)

ConvertFacts(samp.unit)

EstChat(Qdat, modobj, modobjfix)

EstLoad(
  CQdat,
  Qdat,
  Xdesign,
  W,
  sig2,
  rho,
  alpha1,
  alpha2,
  beta1,
  samp.unit,
  se = TRUE,
  Blocklen = 5000
)

EstLoadYS(
  CQdat,
  Qdat,
  concY,
  modobj,
  samp.unit = samp.unit,
  alpha1,
  alpha2,
  se = TRUE,
  Blocklen = 5000
)

ExtractDailyLoads(Date, Xdesign, Lhat, LhatC, Qhat, error, modparms)

Intervals(est, pvalue)

Arguments

len
len

Xdesign
Design matrix

W
Var(betahat)

LhatC
Concentration estimate

alpha1
measurement error

alpha2
error in flow rates

beta1
coeficient of log Q in the regression

rho
rho

sig2
sigma^2

samp.unit
sampling unit

Qdat
Qdat

modobj
modobj

modobjfix
modobjfix

CQdat
Predicted concentration and flow

se
se (Default: TRUE)

Blocklen
Block length (Default: 5000)

concY
concY

Date
Date

Lhat
Load estimate

Qhat
Flow estimate

error
error

modparms
modparms

est
estimate

pvalue
pvalue